# docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Character API",
    "version": "3.0"
  },
  "paths": {
    "/single-character": {
      "post": {
        "summary": "Add or update a single character",
        "description": "Add or update a single character",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "character",
            "description": "Character data",
            "required": true,
            "schema": {
              
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Character added or updated successfully",
            "schema": {
              "type": "string"
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "get": {
        "summary": "Get a single character",
        "description": "Retrieve a single character by token",
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "query",
            "name": "character_token",
            "description": "Character token",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Character retrieved successfully",
            "schema": {
              }
          },
          "400": {
            "description": "Invalid request"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/characters": {
      "post": {
        "summary": "Add multiple characters",
        "description": "Add multiple characters in bulk",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "characters",
            "description": "List of characters to add",
            "required": true,
            "schema": {
              "type": "array",
              "items": {
                
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Characters added successfully",
            "schema": {
              "type": "string"
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "get": {
        "summary": "Get all characters",
        "description": "Retrieve all characters",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "Characters retrieved successfully",
            "schema": {
              "type": "array",
              "items": {
                
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/new-encounter": {
      "post": {
        "summary": "Reset initiative for a new encounter",
        "description": "Reset initiative for a new encounter",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "Initiative reset successfully",
            "schema": {
              "type": "string"
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/next-combatant": {
      "get": {
        "summary": "Get next combatant in initiative order",
        "description": "Get next combatant in initiative order",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "Next combatant retrieved successfully",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    }
  }
}

    